refactor: resolve code size limits by modularizing db, backup-validator, google-drive, and export-utils layers
- Created src/lib/db/ for importing and merging transactions.
- Created src/lib/backup-validator/ for domain validations.
- Created src/lib/google-drive/ for API operations, auth, and HTTP clients.
- Created src/lib/export/ for canvas drawing steps.
- Ensured all functions are under 50 lines and files are under 300 lines.
- Updated Memory Bank documents (systemPatterns.md, techContext.md).
- Verified all tests (446 passed) and ESLint rules (0 errors).
- This completes the refactoring requirements for Issue #440.
